ShapeManager. Правим баг когда шейп без текста нельзя уменьшить до конца путём пропорционального скейлинга по диагонали
