+++ e2e/fixtures/editor.fixture.ts
@@ -2,7 +2,7 @@ import { readFile } from 'node:fs/promises'
 import { setTimeout as delay } from 'node:timers/promises'
 import { test as base } from '@playwright/test'
 import { EditorModel } from '../models/editor.model'
-import { ShapeModel } from '../models/shape.model'
+import { ShapeModel } from '../models/shape/shape.model'
 import { CanvasModel } from '../models/canvas.model'
 import { HistoryModel } from '../models/history.model'
 import { ClipboardModel } from '../models/clipboard.model'
