+++ playwright.config.ts
@@ -1,4 +1,9 @@
-import { defineConfig, devices } from '@playwright/test';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { defineConfig, devices } from '@playwright/test'
+
+const PLAYWRIGHT_HOST = '127.0.0.1'
+const PLAYWRIGHT_PORT = Number(process.env.PLAYWRIGHT_PORT ?? 4173)
+const PLAYWRIGHT_BASE_URL = `https://${PLAYWRIGHT_HOST}:${PLAYWRIGHT_PORT}`
 
 /**
  * Read environment variables from file.
@@ -20,22 +25,23 @@ export default defineConfig({
   reporter: 'html',
 
   use: {
-    baseURL: 'https://localhost:5173',
+    baseURL: PLAYWRIGHT_BASE_URL,
     ignoreHTTPSErrors: true,
-    trace: 'on-first-retry',
+    trace: 'on-first-retry'
   },
 
   projects: [
     {
       name: 'chromium',
-      use: { ...devices['Desktop Chrome'] },
-    },
+      use: { ...devices['Desktop Chrome'] }
+    }
   ],
 
+  // E2E always runs against its own Vite server and must not depend on the user's main dev server.
   webServer: {
-    command: 'npm run dev',
-    url: 'https://localhost:5173',
+    command: `npm run dev:e2e -- --port ${PLAYWRIGHT_PORT}`,
+    url: PLAYWRIGHT_BASE_URL,
     ignoreHTTPSErrors: true,
-    reuseExistingServer: !process.env.CI,
-  },
-});
+    reuseExistingServer: !process.env.CI
+  }
+})
