+++ tsconfig.json
@@ -108,9 +108,9 @@
     /* Completeness */
     // "skipDefaultLibCheck": true,                      /* Skip type checking .d.ts files that are included with TypeScript. */
     "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
-    "moduleResolution": "node",                     /* Specify how TypeScript looks up a file from a given module specifier. */
+    "moduleResolution": "bundler",                  /* Use bundler-style resolution for the Vite-based project and avoid deprecated node10 resolution. */
     "allowSyntheticDefaultImports": true             /* Allow 'import x from y' when a module doesn't have a default export. */
   },
-  "include": ["src/**/*", "docs/demo", "jest.config.ts", "specs/**/*", "e2e/**/*"],         // Папки/файлы, включаемые в компиляцию
+  "include": ["src/**/*", "docs/demo", "jest.config.ts", "playwright.config.ts", "specs/**/*", "e2e/**/*"],         // Папки/файлы, включаемые в компиляцию
   "exclude": ["node_modules", "dist"]
 }
